Update climate.markdown
Trough the forum I found out there is a `hvac_action`  attribute, which shows the current state. This wasn't yet available in the documentation. Therefore chose to add this. 

I wasn't sure which page this should be placed, either the climate description or the generic_thermostat. I chose Climate as it is a climate attribute. 

This attribute is described here: https://www.home-assistant.io/blog/2019/07/17/release-96/#climate
This was the forum post which lead me to it:
https://community.home-assistant.io/t/generic-thermostat-state/140174

Here is a forum post which also describes the `fan` attribute. 
https://community.home-assistant.io/t/how-to-trigger-automation-from-hvac-action/202597